Fix showing output for non-sortable sets of values.
When showing output from a failed comparison where sets are involved, we
try to sort the values to provide consistent odering for comparison.
When doing this, we could hit an error saying that the values aren't
sortable, due to `__lt__` not being implemented.

We now detect this case and sort by the representations of the values
instead.

Testing Done:
All unit tests pass.

Verified this fixed issues encountered in some new unit tests in
Review Board.

Reviewed at https://reviews.reviewboard.org/r/14386/