Add release notes for django-assert-queries 2.0.1.
This is a bug fix release that fixes comparisons of sets of models,
equality checks between two of the same instance of a model, and the
display of assertion failure results on pytest.

Testing Done:
Built the docs. Checked for build and spelling errors.

Reviewed at https://reviews.reviewboard.org/r/14388/